# lib/api-zod/src/generated/types/bountyProgram.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S_A-H ULTIMATE — Web3 Audit Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { BountyProgramPlatform } from "./bountyProgramPlatform";

export interface BountyProgram {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  maxBounty?: number | null;
  maxBountyDisplay: string;
  url: string;
  platform: BountyProgramPlatform;
  assets: string[];
  tags: string[];
  active: boolean;
  kycNote: string;
  /** @nullable */
  lastUpdated?: string | null;
}