# lib/api-zod/src/generated/types/harvestJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S_A-H ULTIMATE — Web3 Audit Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { HarvestJobStatus } from "./harvestJobStatus";

export interface HarvestJob {
  jobId: string;
  repoUrl: string;
  status: HarvestJobStatus;
  createdAt: string;
  fileCount: number;
  /** @nullable */
  contextSizeKb?: number | null;
}